Update quickstart.rst
I added articles-model.rst to quick start cms tutorial, to avoid confusion. Leaving one link there and not even saying that it is related to the tutorial but just writing there (With the database built, we can now build `Models`.) is not very convenient for readers. Many will not understand that this is a continuation of the tutorial, especially beginners. Therefore, suggesting a page about creating a model, also add it here with all the others